SwiftFormatFile in its own file
